+++ Engine/Plugins/Runtime/ReplicationSystemTestPlugin/Source/Private/Tests/Serialization/TestNetRoleNetSerializer.cpp
@@ -194,7 +194,7 @@ void FTestNetRoleNetSerializer::TestValidate()
 	{
 		TArray<bool> ExpectedResults;
 		ExpectedResults.SetNumUninitialized(Values.Num());
-		for (SIZE_T ValueIt = 0, ValueEndIt = Values.Num(); ValueIt != ValueEndIt; ++ValueIt)
+		for (int32 ValueIt = 0, ValueEndIt = Values.Num(); ValueIt != ValueEndIt; ++ValueIt)
 		{
 			ExpectedResults[ValueIt] = true;
 		}
